Update for OpenTrack 13.1 Release to Remove ServerName as required.
From the Release Notes:
It will no longer be necessary to enter the ServerName information in the Dealer section of any OpenTrack request.  OpenTrack is going to utilize a table to associate the Dealership’s enterprise code and company number with the appropriate server.  All that will be need to be populated in the Dealer section of the xml is the EnterpriseCode and CompanyNumber to send a request through OpenTrack for the appropriate dealership.